Split admin functional tests into read and write files for AB#16864.
